requirements.txt
flask
gunicorn
PyPDF2
keys
google.generativeai
streamlit